SPECIAL branch - eForth STM32F405
Target this sketch: Trinket M0

The Trinket M0 is used to talk to the STM32F405
board (running eForth) via the two boards'
USART chips (on their TX/RX pairs).

Main purpose of this branch is to account for
subtle differences in communications, based on
upstream eForth author's design parameters for
this particular Forth.

Also, in a recent commit, turned off the Trinket M0's
dotstar RGB LED.

This commit: just a rename of a legacy experiment's
related .ino file (they are to be named as their
directory).

	renamed:    m0_terminal_uart-3nov-a.ino -> m0_terminal_uart.ino

On branch testing-ccg-